Create some more tests for TodoControllerSpec
Finish some more tests that assure that TodoController is actually working correctly. 
NOTE: Both GetTodosByKeyWord, and GetTodosWithIllegalStatus do not currently work. 
It's possible that the IllegalStatus test is accidentally typecasting a string to a boolean. The KeyWord test, however, may be failing due to the actual filtering not being implemented correctly